Refactored code, no change
